requirements.txt
asgiref==3.7.2
astroid==3.1.0
dill==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
tomlkit==0.12.4
uritemplate==4.1.1
